+++ MatrixMulti.cpp
@@ -6,29 +6,29 @@ MatrixMulti::MatrixMulti(InputTable *table, int begin, int end)
     partData = CalcMatrix();
 }
 
-std::vector<std::vector <int> > MatrixMulti::CalcMatrix() 
+std::vector<std::vector <short> > MatrixMulti::CalcMatrix() 
 {
-    std::vector<std::vector <int> > orgMatrix = mainTable->GetRowData();
-    std::vector<std::vector <int> > newMatrix;
-    std::vector<std::vector <int> >::iterator it;
+    std::vector<std::vector <short> > orgMatrix = mainTable->GetRowData();
+    std::vector<std::vector <short> > newMatrix;
+    std::vector<std::vector <short> >::iterator it;
     for(it = orgMatrix.begin(); it < orgMatrix.end(); it++) {
-	std::vector<int> numbers;
-	std::vector<int> tempVector = *it;
+	std::vector<short> numbers;
+	std::vector<short> tempVector = *it;
 	for(int i = frameBegin - 1; i != frameEnd; i++) {
 	    numbers.push_back(tempVector[i]);
 	}
 	newMatrix.push_back(numbers);
     }
-    std::vector<std::vector <int> > adjMatrix;
+    std::vector<std::vector <short> > adjMatrix;
     
     for(it = newMatrix.begin(); it != newMatrix.end(); it++) {
-	std::vector<int> firstRow = *it;
-	std::vector<std::vector <int> >::iterator it2;
-	std::vector<int> newRow;
+	std::vector<short> firstRow = *it;
+	std::vector<std::vector <short> >::iterator it2;
+	std::vector<short> newRow;
 	for(it2 = newMatrix.begin(); it2 != newMatrix.end(); it2++) {
-	    std::vector<int> secondRow = *it2;
+	    std::vector<short> secondRow = *it2;
 	    int sum = 0;
-	    for(std::vector<int>::size_type i = 0; i != (frameEnd - frameBegin + 1); i++) {
+	    for(std::vector<short>::size_type i = 0; i != (frameEnd - frameBegin + 1); i++) {
 		sum = sum + (firstRow[i] * secondRow[i]);
 	    }
 	    newRow.push_back(sum);
@@ -39,7 +39,7 @@ std::vector<std::vector <int> > MatrixMulti::CalcMatrix()
 }
 
 
-const std::vector<std::vector <int> > MatrixMulti::GetPartData() 
+const std::vector<std::vector <short> > MatrixMulti::GetPartData() 
 {
     return partData;
 }
